Use type uint32 instead of float64 for the weights of statistics.

+++ cpp/subprojects/boosting/include/mlrl/boosting/data/vector_statistic_decomposable_dense.hpp
@@ -58,7 +58,7 @@ namespace boosting {
              * @param row       The index of the row to be added to this vector
              * @param weight    The weight, the gradients and Hessians should be multiplied by
              */
-            void add(const CContiguousView<Tuple<float64>>& view, uint32 row, float64 weight);
+            void add(const CContiguousView<Tuple<float64>>& view, uint32 row, uint32 weight);
 
             /**
              * Removes all gradients and Hessians in a single row of a `CContiguousView` from this vector.
@@ -78,7 +78,7 @@ namespace boosting {
              * @param row       The index of the row to be removed from this vector
              * @param weight    The weight, the gradients and Hessians should be multiplied by
              */
-            void remove(const CContiguousView<Tuple<float64>>& view, uint32 row, float64 weight);
+            void remove(const CContiguousView<Tuple<float64>>& view, uint32 row, uint32 weight);
 
             /**
              * Adds certain gradients and Hessians in a single row of a `CContiguousView`, whose positions are given as
@@ -116,7 +116,7 @@ namespace boosting {
              * @param weight    The weight, the gradients and Hessians should be multiplied by
              */
             void addToSubset(const CContiguousView<Tuple<float64>>& view, uint32 row,
-                             const CompleteIndexVector& indices, float64 weight);
+                             const CompleteIndexVector& indices, uint32 weight);
 
             /**
              * Adds certain gradients and Hessians in single row of a `CContiguousView`, whose positions are given as a
@@ -130,7 +130,7 @@ namespace boosting {
              * @param weight    The weight, the gradients and Hessians should be multiplied by
              */
             void addToSubset(const CContiguousView<Tuple<float64>>& view, uint32 row, const PartialIndexVector& indices,
-                             float64 weight);
+                             uint32 weight);
 
             /**
              * Sets the gradients and Hessians in this vector to the difference `first - second` between the gradients
